python-backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
openai
supabase
python-dotenv
httpx[socks]
pytest
